[fix] redis队列消费

+++ g2-starter-parent/g2-starter-core/src/main/java/org/g2/core/thread/scheduler/ScheduledTask.java
@@ -34,7 +34,7 @@ public class ScheduledTask extends TimerTask {
     private final AtomicInteger throwableCounter;
     private final AtomicInteger timeoutCounter;
 
-    public ScheduledTask(String name, ScheduledExecutorService scheduler, ThreadPoolExecutor executor, Runnable task, int delay, TimeUnit timeUnit, boolean timeout) {
+    public ScheduledTask(String name, ScheduledExecutorService scheduler, ThreadPoolExecutor executor, Runnable task, long delay, TimeUnit timeUnit, boolean timeout) {
         this.name = name;
         this.scheduler = scheduler;
         this.executor = executor;
